+++ test/index_test.js
@@ -67,6 +67,12 @@ describe("Range", function() {
     })
   })
 
+  describe(".prototype.toJSON", function() {
+    it("must be an alias to toString", function() {
+      Range.prototype.toJSON.must.equal(Range.prototype.toString)
+    })
+  })
+
   describe(".parse", function() {
     it("must parse string with bounds", function() {
       Range.parse("[a,z]").must.eql(new Range("a", "z", "[]"))